feat(ever-admin): created shared Warehouse Products table EV-1197
