+++ admin/website-angular/src/app/@shared/warehouse-product/forms/warehouse-product-forms.module.ts
@@ -6,6 +6,15 @@ import { WarehouseAddChoiceComponent } from './warehouse-add-choice';
 import { AddWarehouseProductsComponent } from './add-warehouse-products-table';
 import { Ng2SmartTableModule } from 'ng2-smart-table';
 import { RenderComponentsModule } from '../../render-component/render-components.module';
+import { WarehouseProductsComponent } from './warehouse-products-table';
+import { StoreProductsTableModule } from 'app/@shared/render-component/store-products-table/store-products-table.module';
+import { ProductCategoriesModule } from 'app/@shared/render-component/product-categories/product-categories.module';
+
+const COMPONENTS = [
+	WarehouseAddChoiceComponent,
+	AddWarehouseProductsComponent,
+	WarehouseProductsComponent
+];
 
 @NgModule({
 	imports: [
@@ -13,9 +22,11 @@ import { RenderComponentsModule } from '../../render-component/render-components
 		FormWizardModule,
 		TranslateModule.forChild(),
 		Ng2SmartTableModule,
-		RenderComponentsModule
+		RenderComponentsModule,
+		StoreProductsTableModule,
+		ProductCategoriesModule
 	],
-	exports: [WarehouseAddChoiceComponent, AddWarehouseProductsComponent],
-	declarations: [WarehouseAddChoiceComponent, AddWarehouseProductsComponent]
+	exports: COMPONENTS,
+	declarations: COMPONENTS
 })
 export class WarehouseProductFormsModule {}
