+++ admin/website-angular/src/app/@shared/warehouse-product/forms/warehouse-products-table/index.ts
@@ -1,0 +1,1 @@
+export * from './warehouse-products-table.component';
