+++ admin/website-angular/src/app/@shared/warehouse-product/forms/warehouse-products-table/warehouse-products-table.component.html
@@ -1,0 +1,8 @@
+<ng2-smart-table
+	[settings]="settingsSmartTable"
+	[source]="sourceSmartTable"
+	(userRowSelect)="selectProductTmp($event)"
+	(edit)="onEdit.emit($event)"
+	(delete)="onDelete.emit($event)"
+>
+</ng2-smart-table>
